SettingProvider interface
