+++ src/main/kotlin/com/github/noamm9/ui/clickgui/components/Setting.kt
@@ -1,11 +1,7 @@
 package com.github.noamm9.ui.clickgui.components
 
 import com.github.noamm9.NoammAddons
-import com.github.noamm9.features.Feature
-import com.github.noamm9.ui.clickgui.components.impl.CategorySetting
-import com.github.noamm9.ui.clickgui.components.impl.SeparatorSetting
 import net.minecraft.client.gui.GuiGraphicsExtractor
-import kotlin.reflect.KProperty
 
 abstract class Setting<T>(val name: String, val defaultValue: T) {
     open var value: T = defaultValue
@@ -38,53 +34,4 @@ abstract class Setting<T>(val name: String, val defaultValue: T) {
 
     open fun keyPressed(keyCode: Int, scanCode: Int, modifiers: Int) = false
     open fun charTyped(codePoint: Char) = false
-
-    companion object {
-        fun <T: Setting<*>> T.section(name: String): T {
-            this.headerName = name
-            return this
-        }
-
-        fun <T: Setting<*>> T.withDescription(desc: String): T {
-            this.description = desc.let {
-                return@let if (! it.endsWith('.')) "$it."
-                else it
-            }
-            return this
-        }
-
-        fun <T, S: Setting<T>> S.onChange(listener: (T) -> Unit): S {
-            this.changeListener = listener
-            return this
-        }
-
-        fun <T: Setting<*>> T.showIf(condition: () -> Boolean): T {
-            this.visibility = condition
-            return this
-        }
-
-        fun <T: Setting<*>> T.hideIf(condition: () -> Boolean): T {
-            this.visibility = { ! condition() }
-            return this
-        }
-
-
-        operator fun <T, S: Setting<T>> S.provideDelegate(thisRef: Feature, prop: KProperty<*>): S {
-            this.headerName?.let { name ->
-                if (thisRef.configSettings.isNotEmpty()) {
-                    thisRef.configSettings.add(SeparatorSetting().also { it.visibility = this.visibility })
-                }
-                thisRef.configSettings.add(CategorySetting(name).also {
-                    it.visibility = this.visibility
-                })
-            }
-
-            thisRef.configSettings.add(this)
-            return this
-        }
-
-        operator fun <T, S: Setting<T>> S.getValue(thisRef: Feature, prop: KProperty<*>): S {
-            return this
-        }
-    }
 }
